QR_word_to_sound_ru/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
gTTS==2.2.3
idna==3.3
pypng==0.0.21
PyQRCode==1.2.1
python-dotenv==1.0.0
python-telegram-bot==13.11
pytz==2023.3.post1
requests==2.26.0
six==1.16.0
tornado==6.3.3
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.7
